refactor: move JSON-RPC error codes to header
