+++ libtransmission/rpcimpl.h
@@ -11,6 +11,33 @@
 struct tr_session;
 struct tr_variant;
 
+namespace JsonRpc
+{
+auto constexpr Version = std::string_view{ "2.0" };
+
+namespace Error
+{
+enum Code : int16_t
+{
+    PARSE_ERROR = -32700,
+    INVALID_REQUEST = -32600,
+    METHOD_NOT_FOUND = -32601,
+    INVALID_PARAMS = -32602,
+    INTERNAL_ERROR = -32603,
+    SUCCESS = 0,
+    SET_ANNOUNCE_LIST,
+    INVALID_TRACKER_LIST,
+    PATH_NOT_ABSOLUTE,
+    UNRECOGNIZED_INFO,
+    SYSTEM_ERROR,
+    FILE_IDX_OOR,
+    HTTP_ERROR,
+    CORRUPT_TORRENT,
+    N_ERROR
+};
+}
+} // namespace JsonRpc
+
 using tr_rpc_response_func = std::function<void(tr_session* session, tr_variant&& response)>;
 
 void tr_rpc_request_exec(tr_session* session, tr_variant const& request, tr_rpc_response_func&& callback = {});
